+++ tools/tracevis/js/waterfall.js
@@ -15,17 +15,6 @@
  */
 
 var WATERFALL = (function() {
-  var _traceClass = function (d) {
-    var cls = ["trace", d.resultType];
-    if (d.children || d._children) {
-      cls.push("composite");
-      if (d._children) {
-        cls.push("collapsed");
-      }
-    }
-    return cls.join(' ');
-  };
-
   var _toMillisFloat = function(time) {
     return time / 1000000;
   };
@@ -33,13 +22,10 @@ var WATERFALL = (function() {
   /**
    * Renders the view and append it to a d3 selection.
    */
-  var render = function(selection, root) {
+  function render(selection, root) {
     selection.classed("waterfallview", true);
 
     var w = 960,
-
-        // We update height when we know how tall we need the chart
-        h = 0,
 
         // We use a margin for the axes
         margin = { top: 40, bottom: 10, left: 10, right: 250 },
@@ -48,43 +34,68 @@ var WATERFALL = (function() {
         minBarWidth = 3,
 
         // transition duration
-        duration = 0,
+        defaultDuration = 250,
 
-        // root of the tree
-        root;
+        alphaDimmed = 0.3,
+
+        taskFill = { "default":       "#DDD",
+                     "success":       "#E0FFE0",
+                     "error":         "#FFE0E0",
+                     "early_finish":  "#FFC" };
+
+    // Mapping of descendant to a set (expressed as an object) of a ancestors
+    var ancestors = {};
+    function ancestorPreorder(x) {
+      var xAncestors = ancestors[x.id] = {};
+      if (x.parent) {
+        xAncestors[x.parent.id] = true;
+        for (var ancestorId in ancestors[x.parent.id]) {
+          xAncestors[ancestorId] = true;
+        }
+      }
+      if (x.children) {
+        for (var i = 0; i < x.children.length; ++i) {
+          ancestorPreorder(x.children[i]);
+        }
+      }
+    }
+    ancestorPreorder(root)
 
     var svg = selection.append("svg")
       .attr("width", w + margin.left + margin.right);
 
+    var vis = svg.append("g")
+      .attr("transform", "translate(" + margin.left + "," + margin.top +")")
+      .append("g");
+
     var xLabel = svg.append("g")
       .attr("class", "x axis label")
-      .attr("transform", "translate(" + margin.left + "," + (margin.top - 15) + ")");
+      .attr("transform", "translate(" + margin.left + "," + (margin.top - 15) + ")")
+      .append("g");
 
     var xGrid = svg.append("g")
       .attr("class", "x axis grid")
-      .attr("transform", "translate(" + margin.left + "," + (margin.top - 15) + ")");
+      .attr("transform", "translate(" + margin.left + "," + (margin.top - 15) + ")")
+      .append("g");
 
-    var vis = svg.append("g")
-      .attr("transform", "translate(" + margin.left + "," + margin.top +")");
+    update(root, 0);
 
-    function update(source) {
+    function update(source, duration) {
       var traces = TRACE.flatten(root);
 
-      // Resize the canvas
-      var newHeight = traces.length * (barHeight + barSpacing) + margin.top + margin.bottom;
-      if (newHeight > h) {
-        h = newHeight;
-        svg.attr("height", h);
-      }
-
-      // Compute the "layout".
-      var xScale = d3.scale.linear()
+      var x = d3.scale.linear()
         .domain([0, d3.max(traces, function(d) { return d.start + d.elapsed; })])
         .range([0, w]);
 
+      // Resize the canvas
+      var h = traces.length * (barHeight + barSpacing) + margin.top + margin.bottom;
+      if (svg.attr("height") - h < 0) {
+        svg.attr("height", h);
+      }
+
       traces.forEach(function(d, i) {
-        d.x = xScale(_toMillisFloat(d.startNanos));
-        d.w = xScale(_toMillisFloat(d.startNanos + d.elapsedNanos)) - d.x;
+        d.x = x(_toMillisFloat(d.startNanos));
+        d.w = x(_toMillisFloat(d.startNanos + d.elapsedNanos)) - d.x;
 
         if (d.w < minBarWidth) {
           d.x = Math.max(d.x - minBarWidth, d.parent ? d.parent.x : 0);
@@ -95,67 +106,121 @@ var WATERFALL = (function() {
         d.h = barHeight;
       });
 
-      xGrid.call(d3.svg.axis()
-        .scale(xScale)
+      xGrid.transition()
+        .duration(duration)
+        .call(d3.svg.axis()
+        .scale(x)
         .orient("top")
         .tickSubdivide(10)
         .tickSize(-h,-h,0)
         .tickFormat(""));
 
       xLabel.call(d3.svg.axis()
-        .scale(xScale)
+        .scale(x)
         .orient("top")
         .tickSubdivide(10)
         .tickSize(6, 3, 0));
 
       // Update the traces
       var trace = vis.selectAll("g.trace")
-        .data(traces, function(d) { return d.id })
-        .attr("class", _traceClass);
+        .data(traces, function(d) { return d.id; });
 
       var traceEnter = trace.enter().append("g")
-        .attr("class", _traceClass)
+        .classed("trace", true)
+        .classed("composite", function(d) { return d.children || d._children; })
+        .classed("hidden", false)
         .attr("transform", function(d) { return "translate(" + source.x0 + "," + source.y0 + ")"; })
-        .style("opacity", 1e-6);
+        .style("opacity", 1e-6)
+        .on("click", toggleCollapse);
+
+      trace.classed("collapsed", function(d) { return d._children; });
 
       // Enter any new traces at the parent's previous position.
       traceEnter.append("rect")
         .attr("y", -barHeight / 2)
+        .attr("rx", 3)
         .attr("height", function(d) { return d.h; })
-        .attr("width", function(d) { return d.w; });
+        .attr("width", function(d) { return d.w; })
+        .style("fill", function(d) { return taskFill[d.resultType] || taskFill["default"]; })
+        .on("mouseover", mouseover)
+        .on("mouseout", mouseout);
 
-      traceEnter.append("text")
+      var textEnter = traceEnter.append("text")
         .attr("dy", 4)
-        .attr("dx", 6)
-        .text(function(d) { return d.name + " (" + d.elapsed + " ms)"; });
+        .attr("dx", 6);
 
-      // Transition traces to their new position.
-      traceEnter.transition()
-        .duration(duration)
-        .attr("transform", function(d) { return "translate(" + d.x + "," + d.y + ")"; })
-        .style("opacity", 1);
+      textEnter.append("tspan").attr("class", "expando");
+      textEnter.append("tspan").text(function(d) { return d.name + " (" + d.elapsed + " ms)"; });
+
+      var text = trace.select("text tspan.expando")
+          .style("font-family", "monospace")
+          .text(function(d) {
+              if (d.children) {
+                return "[-] ";
+              } else if (d._children) {
+                return "[+] ";
+              }
+            });
 
       trace.transition()
         .duration(duration)
         .attr("transform", function(d) { return "translate(" + d.x + "," + d.y + ")"; })
-        .style("opacity", 1);
+          .style("opacity", function(d) { return d.waterfallDimmed ? alphaDimmed : 1; })
+          .each("end", function() { d3.select(this).style("opacity", function(d) { return d.waterfallDimmed ? alphaDimmed : 1; }); });
+
+      // Transition traces to their new position.
+      traceEnter
+        .style("pointer-events", "none")
+        .transition()
+          .duration(duration)
+          .attr("transform", function(d) { return "translate(" + d.x + "," + d.y + ")"; })
+          .style("opacity", function(d) { return d.waterfallDimmed ? alphaDimmed : 1; })
+          .each("end", function() { d3.select(this).style("pointer-events", undefined); });
 
       // Transition exiting traces to the parent's new position.
-      trace.exit().transition()
-        .duration(duration)
-        .attr("transform", function(d) { return "translate(" + source.x + "," + source.y + ")"; })
-        .style("opacity", 1e-6)
-        .remove();
+      trace.exit()
+        .style("pointer-events", "none")
+        .classed("hidden", true)
+        .transition()
+          .duration(duration)
+          .attr("transform", function(d) { return "translate(" + source.x + "," + source.y + ")"; })
+          .style("opacity", 1e-6)
+          .each("end", function() { d3.select(this).style("pointer-events", undefined); })
+          .remove();
 
       // Stash the old positions for transition.
       traces.forEach(function(d) {
         d.x0 = d.x;
         d.y0 = d.y;
       });
-
     }
 
-    update(root);
+    function toggleCollapse(d) {
+      if (d.children) {
+        d._children = d.children;
+        delete d.children;
+      } else {
+        d.children = d._children;
+        delete d._children;
+      }
+      update(d, defaultDuration);
+    }
+
+    function mouseover(d) {
+      if (d.children || d._children) {
+        vis.selectAll("g.trace")
+          .filter(function(e) { return e !== d && !ancestors[e.id][d.id]; })
+          .each(function(e) { e.waterfallDimmed = true; })
+          .style("opacity", alphaDimmed);
+      }
+    }
+
+    function mouseout(d) {
+      vis.selectAll("g.trace")
+        .filter(function(e) { return e.waterfallDimmed; })
+        .style("opacity", function() { return d3.select(this).classed("hidden") ? 0 : 1 })
+        .each(function(e) { delete e.waterfallDimmed; });
+    }
   };
 
   return {